Create FA3.py
First, we have total_sum in 0, and we ask the user to input their age. The i in range(1, age +1) is for it to list down the numbers from 1 all the way to the age given and the “i” will take on each number from the range. The loop will keep on running for each number, and it will get added to total_sum. Last, the print part simply outputs the sum of all the numbers from 1 up to the given age, with the commas separating the variables and words.