Create FA4.py
We started with the input to get the number of our students, subjects, and then for the scores. The variables keep track of each student’s total score, the total scores for the class, and counts how many scores entered. The for loop will allow this to repeat the labels such as “Student 1, Student 2..” and ask for the scores per subject. It will ask to input the scores, and then it calculates for the average. Lastly, it prints the average scores per student, and the average for the class.